Refatoração de processo de registro de rotas.
