+++ infraestrutura/roteador/metodo-roteador.ts
@@ -1,0 +1,14 @@
+export const MetodoRoteador = {
+    Get: 'Get',
+    Post: 'Post',
+    Put: 'Put',
+    Delete: 'Delete',
+    Patch: 'Patch',
+    Options: 'Options',
+    Copy: 'Copy', 
+    Head: 'Head',
+    Lock: 'Lock',
+    Unlock: 'Unlock',
+    Purge: 'Purge',
+    Propfind: 'Propfind'
+};
