+++ interfaces/interface-liquido.ts
@@ -25,16 +25,5 @@ export interface LiquidoInterface {
     prepararRequisicao(requisicao: any, nomeFuncao: string, funcaoConstruto: FuncaoConstruto): void;
 
     chamarInterpretador(nomeFuncao: string): Promise<RetornoInterpretador>;
-    adicionarRotaGet(caminhoRota: string, argumentos: Construto[]): void;
-    adicionarRotaPost(caminhoRota: string, argumentos: Construto[]): void;
-    adicionarRotaPut(caminhoRota: string, argumentos: Construto[]): void;
-    adicionarRotaDelete(caminhoRota: string, argumentos: Construto[]): void;
-    adicionarRotaPatch(caminhoRota: string, argumentos: Construto[]): void;
-    adicionarRotaOptions(caminhoRota: string, argumentos: Construto[]): void;
-    adicionarRotaCopy(caminhoRota: string, argumentos: Construto[]): void;
-    adicionarRotaHead(caminhoRota: string, argumentos: Construto[]): void;
-    adicionarRotaLock(caminhoRota: string, argumentos: Construto[]): void;
-    adicionarRotaUnlock(caminhoRota: string, argumentos: Construto[]): void;
-    adicionarRotaPurge(caminhoRota: string, argumentos: Construto[]): void;
-    adicionarRotaPropfind(caminhoRota: string, argumentos: Construto[]): void;
+    adicionarRota(metodoRoteador: string, caminhoRota: string, argumentos: Construto[]): void;
 }
